requirements.txt
streamlit==1.39.0
pandas==2.2.2
requests==2.31.0 
groq==0.9.0 
langchain==0.2.5
langchain-core==0.2.38
langchain-groq==0.1.6
langchain-community==0.2.5
langchain-text-splitters==0.2.4 
dataclasses-json==0.6.7 
langsmith==0.1.147 
marshmallow==3.26.1 
mypy-extensions==1.1.0 
typing-inspect==0.9.0 
numpy==1.26.4 
altair==5.3.0 
pydantic==2.7.1 
python-dotenv==1.0.1 
aiohttp==3.9.5 
tqdm==4.66.4
langchainhub==0.1.20
plotly==5.22.0
python-dotenv==1.0.1
tavily-python==0.3.3
filelock==3.15.4
fsspec==2024.6.1   # <-- ADDED: For transformers & torch
scipy==1.14.0       # <-- ADDED: For sentence-transformers
networkx==3.3       # <-- ADDED: For torch
joblib==1.4.2       # <-- ADDED: For scikit-learn
threadpoolctl==3.5.0 # <-- ADDED: For scikit-learn
httpx>=0.27.0  # Force a modern version to handle proxy settings correctly
newsapi-python==0.2.7